requirements.txt
# Chess Analysis Dependencies
python-chess==1.999
stockfish==3.28.0
supabase>=2.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pandas>=2.1.3
numpy>=1.25.2
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
